Adapt to Python 3 dict changes
dict behavior has changed significantly in Python 3. According to PEP
3106, dict.keys(), dict.values() and dict.items() have changed to
provide a set-like, "live" object, instead of a static list. Their
behavior resembles closely that of dict.iterkeys() et al from Python 2.

In order for the code not to break, we need to make sure that all places
relying on the result of the above methods being a list are adapted to
explicitly request that. Luckily, most of the uses of .items() and
.values() should have actually been .iteritems() and .itervalues(), so
they can stay. There are a few cases that need to be updated though.

While at it, drop explicit .keys() calls when iterating over dicts to
make the code cleaner.

Signed-off-by: Apollon Oikonomopoulos <apoikos@dmesg.gr>