Make TestLUClusterVerifyGroupVerifyNodeOs more robust
TestLUClusterVerifyGroupVerifyNodeOs checks the GroupVerifyNodeOS LU log
for the presence of specific messages, one of them being an error about
missing OSes. There are two OSes missing and the test assumes the order
they will be reported, but since the list of missing OSes is constructed
by the difference of sets of dict keys, the order is not guaranteed.
Relax the regex to check for the presence of the canary OS in any place.

Signed-off-by: Apollon Oikonomopoulos <apoikos@dmesg.gr>