utils.text: handle Python 3 type changes
Now that str is unicode and vice versa, SafeEncode must be adjusted not
to try to work with bytes internally. Change the function to always
encode its input and then decode it back from ascii to a string, so that
the rest of the code works.

Additionally, FormatUnit was checking for long, which is gone in Python
3. Rather than just remove long from the allowed types, replace the
whole instance check with numbers.Real, which provides division - the
only property required for this function to work. This will make the
code more robust and future-proof.

Signed-off-by: Apollon Oikonomopoulos <apoikos@dmesg.gr>