serializer: always DumpJson as bytes, load as both
DumpJson is used to prepare objects either for RPC/IPC, or for storage.
In both cases, they should be encoded as bytes, since the data will be
leaving the process boundaries. Adapt DumpJson to always dump objects as
UTF-8-encoded bytes.

On the other hand, we can be more liberal with LoadJson and allow it to
load both, since simplejson handles bytes natively just fime. However,
rename `txt` to `data` to better reflect reality.

Finally, make minor changes to the serializer tests to account for the
API changes.

Signed-off-by: Apollon Oikonomopoulos <apoikos@dmesg.gr>