Convert map to lists where needed
In Python 3, map returns an iterator object instead of a list. This is
okay in some cases, but there are code paths that might break, either
because we explicitly expect a list, or because we e.g. iterate twice
over the result of map.

2to3 converts (almost) all map calls to list(map(…)). Rather than do
that, convert only the calls that need to be converted, and convert
them to list comprehensions, which are more pythonic and readable.

Signed-off-by: Apollon Oikonomopoulos <apoikos@dmesg.gr>