RunCmd: use Popen's pass_fds
Python 3.4 and later have changed the file descriptor inheritance
semantics: file descriptors must be explicitly marked as inheritable to
survive an exec() (see PEP 446[1] for more information). This means that
our noclose_fds handling in RunCmd is currently broken: the FDs survive
the initial fork and subprocess setup, but are closed on exec().

One way to solve this would be to modify our preexec_fn to mark all
noclose_fds as inheritable in the child process before the exec()
happens. However, Python 3.2 offers a more convenient interface, via the
pass_fds argument to Popen, that already does what our noclose_fds wants
to achieve in the first place, so use this and get rid of preexec_fn
altogether.

Note that while the documentation on pass_fds is not very specific
regarding inheritance, the Python subprocess code explicitly marks all
pass_fds as inheritable. Also note that we don't need to touch close_fds
anymore, as pass_fds does the right thing internally.

[1] https://www.python.org/dev/peps/pep-0446/

Signed-off-by: Apollon Oikonomopoulos <apoikos@dmesg.gr>