testutils_ssh: sort master candidate list
The SSH key management tests make implicit assumptions about the order
in which testutils_ssh.FakeSshFileManager returns master candidates: in
particular, they seem to assume that the first 3 master candidates
returned by GetAllMasterCandidates will not include the master node.

Up until Python 3.6 this would work due to the way our test nodes are
called and the fact that dict.items() would return items in *key hash*
order. However, since Python 3.6, dictionaries are ordered in insertion
order and this causes the tests to break.

This is arguably code smell, as the test result should not rely on
implementation details, such as the dict order, however this currently
requires a significant refactoring of the SSH tests. As a stopgap,
change GetAllMasterCandidates to always return the master candidate list
with the master node as the last node. This should make things work
again for the time being.

Signed-off-by: Apollon Oikonomopoulos <apoikos@dmesg.gr>