_RunCmdPipe: close child pipes when done
The pipe(2)s used for capturing the child process stdout/stderr are
allocated by subprocess.Popen() internally and must be explicitly closed
to avoid FD leaks. Close them right when we are done reading from them
in the poll loop, either because there is no data left, or because we
are breaking out of the loop due to a timeout.

Signed-off-by: Apollon Oikonomopoulos <apoikos@dmesg.gr>