Fix Python daemon signal handling with Python 3.5+
daemon.Mainloop relies implicitly for its signal handling on the fact
that asyncore.loop exits when interrupted by a signal. With the advent
of PEP 475 and its implementation in Python 3.5, this is no longer the
case: select.select() and select.poll() will be restarted on EINTR and
the asyncore loop will *not* exit when the process is signaled. This
unfortunately breaks signal handling in noded, failing to process
SIGCHLD and SIGTERM.

Fix this by using an instance of daemon.AsyncAwaker in daemon.Mainloop
and have the signal handlers wake up the asyncore loop explicitly. To
break the chicken-and-egg problem and allow the mainloop to specify its
own waker after having been decorated using utils.HandledSignal, we
extend the API of utils.SignalHandler to allow setting the signal
handling function later.

Signed-off-by: Apollon Oikonomopoulos <apoikos@dmesg.gr>