Fix opening /dev/tty in text mode under Python 3
With the new I/O model of Python 3[1], non-seekable streams such as
/dev/tty cannot be opened read-write in text mode[2]. There is no
universal consensus regarding the workaround, so use io.BufferedRWPair
and io.TextIOWrapper to provide a line-buffered, text-mode IO object for
use. Since this will likely be useful in other places as well, factor it
out to a new utils.io.OpenTTY function.

[1] https://www.python.org/dev/peps/pep-3116/
[2] https://bugs.python.org/issue20074

Signed-off-by: Apollon Oikonomopoulos <apoikos@dmesg.gr>